pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aio-pika==9.6.1",
    "aiormq==6.9.3",
    "alembic==1.18.4",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "asyncpg==0.31.0",
    "bcrypt==5.0.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "ecdsa==0.19.1",
    "fastapi==0.135.1",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "idna==3.11",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "multidict==6.7.1",
    "pamqp==3.3.0",
    "pika==1.3.2",
    "propcache==0.4.1",
    "pyasn1==0.6.2",
    "pydantic[email]==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "python-dotenv==1.2.2",
    "python-jose==3.5.0",
    "python-multipart==0.0.22",
    "rsa==4.9.1",
    "six==1.17.0",
    "sqlalchemy==2.0.48",
    "starlette==0.52.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.41.0",
    "yarl==1.23.0",
    "email-validator>=2.3.0",
    "pwdlib>=0.3.0",
    "redis>=7.3.0",
    "httpx>=0.28.1",
    "python-slugify>=8.0.4",
]
